evolutionary_cot_vlm/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep Learning & Model dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
sentencepiece>=0.1.99
datasets>=2.12.0
Pillow>=9.0.0  # For image processing
requests>=2.28.0  # For downloading data/images

# Evaluation metrics
rouge-score>=0.1.2  # For MMMU evaluation

# Logging and progress bars
tqdm>=4.65.0
logging>=0.5.1.2

# Optional but recommended
jupyter>=1.0.0  # For notebook exploration
tensorboard>=2.12.0  # For monitoring training
scikit-learn>=1.0.2  # For metric calculations 